M5TimerCam/web_cam/bt/requirements.txt
bleak>=0.21.0
Pillow>=10.0.0

